fetch the signers' transactions from the mempool to get the last fees
